Création de l'admin, fO Delete User, Responsive

+++ Back/controller/post.js
@@ -1,11 +1,14 @@
 "use strict";
 
+//FileSystem gestion
 const fs = require('fs');
+//Liens avec les tables
 const db = require("../config/config.js");
 const Post = db.post;
 const User = db.user;
 const Comment = db.comment;
 const Emotions = db.emotions;
+//Middleware : récupérer le userId
 const UserID = require("../middleware/getUserId.js");
 
 //__________________PAGINATION________________________________
@@ -38,8 +41,8 @@ exports.create = (req, res) => {
   let urlGif;
   let urlFile;
   //Si les trois éléments sont nuls alors Erreur 400
-  if( !req.body.textPost && !req.body.gifPost && !req.file ){
-    return res.status(400).json({ message: "Vous devez remplir un élément pour créer une publication." });
+  if( req.body.textPost == '' && req.body.gifPost == '' && req.file == null ){
+    return res.status(400).json('Remplir au moins un élément.');
   }
 
   //Si le text existe création du contenu sinon = NULL
@@ -66,22 +69,20 @@ exports.create = (req, res) => {
 
   // Sauvegarde la publication dans la DB
   Post.create(post)
-    .then((data) => {
-      res.send(data);
+    .then(() => {
+      return res.status(201).json('Nouvelle publication créée !');
     })
     .catch((err) => {
-      res.status(500).send({
-        message: err.message || "Il y a eu une erreur lors de la publication.",
-      });
+      return res.status(500).json(err.message);
     });
 };
 
 //PUT Modifier Publication
 exports.modify = (req, res) => {  
   //Si les trois éléments sont nuls alors Erreur 400
-  if( req.body.textPost == null && req.body.gifPost == null && req.file == null){
-    return res.status(400).json({ message: "Vous devez remplir un élément pour modifier une publication." });
-  } 
+  if( req.body.textPost == '' && req.body.gifPost == '' && req.file == null ){
+    return res.status(400).json('Remplir au moins un élément.');
+  }
 
   const idPost = req.params.id;
   const user = UserID(req);
@@ -91,7 +92,7 @@ exports.modify = (req, res) => {
   Post.findOne({ where : { id : idPost }})
     .then((publication) => {
         if(publication.idUserPost != user) {
-          return res.status(401).send({message: "Vous devez vous identifier."});
+          return res.status(403).json("Vous n'êtes pas le créateur de cette publication.");
         }
 
         //Supprimer l'image de ma publication si ce n'est pas null et mettre null gifPost
@@ -115,42 +116,58 @@ exports.modify = (req, res) => {
         }
       
     })
-  .then(() => {res.status(200).send({message : "Mise à jour de la publication"});})
-  .catch( error => res.status(400).json({ message : "La modification de la publication n'a pas été prise en compte." }, error));
+  .then(() => {res.status(200).json("Mise à jour de la publication");})
+  .catch( () => res.status(400).json("La modification de la publication n'a pas été prise en compte."));
 }
 
 //DELETE Supprimer la publication avec ses commentaires
 exports.delete = (req, res) => {
   const idPost = req.params.id;
+  const user = UserID(req);
   let publication;
+  let ADMIN;
 
-  Post.findOne({ where : { id : idPost }})
-  .then((post) => {
-    publication = post;
-    return Comment.destroy({ where : { postId : publication.id }})
-  })
-  .then(() => {
-    return Emotions.destroy({ where : { idPublication : publication.id }});
-  })
-  .then(() => {
-    if(publication.imagePost){
-      const filename = publication.imagePost.split('/upload/')[1];
-      fs.unlink(`./upload/${filename}`, (error) => {
-        if(error){
-          return console.log(error);
-        }else {
-          return console.log({ message : "Image supprimée"});
-        }
-      })
+  User.findOne({ where : { id : user}})
+    .then((datas) => { 
+        if( datas.isAdmin == 1){ 
+            ADMIN = true; 
+        } else {
+            ADMIN = false;
+        } 
+    })
+    .then(() => {
+      return Post.findOne({ where : { id : idPost }});
+    })
+    .then((post) => {
+      publication = post;
+      if (publication.idUserPost == user || ADMIN ){
+        return Comment.destroy({ where : { postId : publication.id }});
+    }else{
+        return res.status(403).json("Vous n'êtes pas le créateur de cette publication.");
     }
-    return Post.destroy({ where : { id : publication.id }});
-  })
-  .then(() => {
-    res.status(200).send( { message : "La publication est bien supprimée. "});
-  })
-  .catch( error => {
-    res.status(500).send(error);
-  })
+    })
+    .then(() => {
+      return Emotions.destroy({ where : { idPublication : publication.id }});
+    })
+    .then(() => {
+      if(publication.imagePost){
+        const filename = publication.imagePost.split('/upload/')[1];
+        fs.unlink(`./upload/${filename}`, (error) => {
+          if(error){
+            return console.log(error);
+          }else {
+            return console.log({ message : "Image supprimée"});
+          }
+        })
+      }
+      return Post.destroy({ where : { id : publication.id }});
+    })
+    .then(() => {
+      res.status(200).json("La publication est bien supprimée. ");
+    })
+    .catch( error => {
+      res.status(500).send(error);
+    })
 }
 
 //Récupérer toutes les publications
@@ -160,37 +177,55 @@ exports.findAll = (req, res) => {
   const { limit, offset } = getPagination(page, size);
   const createdAt = req.query.createdAt;
   const userId = UserID(req);
+  let ADMIN;
 
-  Post.findAndCountAll({
-    where: createdAt,
-    limit,
-    offset,
-    order: [["createdAt", "DESC"]],
-    include: [ {
-      model : User,
-      as : "UserPublications",
-      attributes : [ "firstname","lastname","imageUrl"] }]
+  User.findOne({ where : { id : userId}})
+    .then((datas) => { 
+        if( datas.isAdmin == 1){ 
+            ADMIN = true; 
+        } else {
+            ADMIN = false;
+        } 
+    })
+  .then(() => {
+    return Post.findAndCountAll({
+      where: createdAt,
+      limit,
+      offset,
+      order: [["createdAt", "DESC"]],
+      include: [ {
+        model : User,
+        as : "UserPublications",
+        attributes : [ "firstname","lastname","imageUrl"] }]
+    });
   })
-    .then((posts) => {
+    .then((posts) => { 
+      if(posts.length == 0){
+        return res.status(200).send(posts);
+      }
       const tabPosts = posts.rows;
       for(let i = 0 ; i < tabPosts.length ; i++){
-        if(userId === tabPosts[i].idUserPost){
-            tabPosts[i].actualUserOK = 1;
-          
-        } else {
+        if(userId === tabPosts[i].idUserPost) {
+          tabPosts[i].actualUserOK = 1;
+        }else {
           tabPosts[i].actualUserOK = 0;
+        }
+        if(ADMIN){
+          tabPosts[i].permission = 1;
+        } else {
+          tabPosts[i].permission = 0;
         }
       }
 
       const response = getPagingData(posts, page, limit);
-      res.status(200).send(response);
+      return res.status(200).send(response);
     })
     .catch((err) => {
-      res.status(500).send({ message: err.message || "Erreur survenue" });
+      return res.status(500).send({ message: err.message });
     });
 };
 
-//Récupérer toutes les publications
+//Récupérer toutes les publications CurrentProfil
 exports.findCurrent = (req, res) => {
   Post.findAll({
     where: { userId: UserID(req) },
@@ -213,10 +248,10 @@ exports.findCurrent = (req, res) => {
     ]
   })
     .then((posts) => {
-      res.send(posts);
+      return res.status(200).send(posts);
     })
-    .catch((error) => {
-      res.status(404).send({ message: "Aucun post trouvé" }, error);
+    .catch(() => {
+      return res.status(404).json("Aucun post trouvé");
     });
 };
 
@@ -226,7 +261,7 @@ exports.findOnePost = (req, res) => {
 
   Post.findOne({ where : { id : id }, attributes : ["imagePost","gifPost","textPost"]})
   .then((publication) => {
-    res.send(publication);
+    return res.status(200).send(publication);
   })
-  .catch((error) => { res.status(404).send({ message: "Les informations ne sont pas retrouvées."}, error)});
+  .catch(() => { return res.status(404).jspn("Les informations ne sont pas retrouvées.")});
 }
